+++ packages/functions/src/lib/initServices.func.ts
@@ -17,7 +17,7 @@ import {parseAccount, parseAccountId} from '@shared/parsers/accounts.parser';
 import {parseAccountSettings} from '@shared/parsers/accountSettings.parser';
 import {parseEventId, parseEventLogItem} from '@shared/parsers/eventLog.parser';
 import {parseAccountExperimentsState} from '@shared/parsers/experiments.parser';
-import {parseFeedItem, parseFeedItemId, toStorageFeedItem} from '@shared/parsers/feedItems.parser';
+import {parseFeedItem, parseFeedItemId} from '@shared/parsers/feedItems.parser';
 import {
   parseUserFeedSubscription,
   parseUserFeedSubscriptionId,
@@ -31,6 +31,7 @@ import {toStorageAccount} from '@shared/storage/accounts.storage';
 import {toStorageAccountSettings} from '@shared/storage/accountSettings.storage';
 import {toStorageEventLogItem} from '@shared/storage/eventLog.storage';
 import {toStorageAccountExperimentsState} from '@shared/storage/experiments.storage';
+import {toStorageFeedItem} from '@shared/storage/feedItems.storage';
 import {toStorageUserFeedSubscription} from '@shared/storage/userFeedSubscriptions.storage';
 
 import {ServerAccountsService} from '@sharedServer/services/accounts.server';
