+++ packages/scripts/src/lib/initServices.scripts.ts
@@ -7,11 +7,12 @@ import {
 } from '@shared/lib/constants.shared';
 
 import {parseEventId, parseEventLogItem} from '@shared/parsers/eventLog.parser';
-import {parseFeedItem, parseFeedItemId, toStorageFeedItem} from '@shared/parsers/feedItems.parser';
+import {parseFeedItem, parseFeedItemId} from '@shared/parsers/feedItems.parser';
 
 import {Environment} from '@shared/types/environment.types';
 
 import {toStorageEventLogItem} from '@shared/storage/eventLog.storage';
+import {toStorageFeedItem} from '@shared/storage/feedItems.storage';
 
 import {ServerEventLogService} from '@sharedServer/services/eventLog.server';
 import {ServerFeedItemsService} from '@sharedServer/services/feedItems.server';
