+++ packages/shared/src/schemas/feedSources.schema.ts
@@ -9,23 +9,29 @@ const BaseFeedSourceSchema = z.object({
   feedSourceType: z.nativeEnum(FeedSourceType),
 });
 
-export const RssFeedSourceSchema = BaseFeedSourceSchema.extend({
+const RssFeedSourceSchema = BaseFeedSourceSchema.extend({
   feedSourceType: z.literal(FeedSourceType.RSS),
   userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
   url: z.string().url(),
   title: z.string(),
 });
 
-export const YouTubeChannelFeedSourceSchema = BaseFeedSourceSchema.extend({
+export type RssFeedSourceFromStorage = z.infer<typeof RssFeedSourceSchema>;
+
+const YouTubeChannelFeedSourceSchema = BaseFeedSourceSchema.extend({
   feedSourceType: z.literal(FeedSourceType.YouTubeChannel),
   userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
   channelId: YouTubeChannelIdSchema,
 });
 
+export type YouTubeChannelFeedSourceFromStorage = z.infer<typeof YouTubeChannelFeedSourceSchema>;
+
 export const IntervalFeedSourceSchema = BaseFeedSourceSchema.extend({
   feedSourceType: z.literal(FeedSourceType.Interval),
   userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
 });
+
+export type IntervalFeedSourceFromStorage = z.infer<typeof IntervalFeedSourceSchema>;
 
 const PwaFeedSourceSchema = BaseFeedSourceSchema.extend({
   feedSourceType: z.literal(FeedSourceType.PWA),
@@ -47,6 +53,8 @@ export const FeedSourceSchema = z.discriminatedUnion('feedSourceType', [
   ExtensionFeedSourceSchema,
   PocketExportFeedSourceSchema,
 ]);
+
+export type FeedSourceFromStorage = z.infer<typeof FeedSourceSchema>;
 
 export const FeedSourceWithUrlSchema = z.discriminatedUnion('feedSourceType', [
   RssFeedSourceSchema,
